chore: add supported builtins config
